Mer refaktorering av enhetstester for driftsmelding
Splitter opp i to filer og pakker dem inn i describe()
